requirements.txt
playwright
pyotp
python-dotenv
requests